build_testraster

FUNCTION_DESCRIPTION

build_testraster(nrows = 100, ncols = 100, nbands = 1,
  with_nodata = TRUE, tofile = FALSE)

Arguments

nrows
numeric number of rows, Default: 100

ncols
numeric number of columns, Default: 100

nbands
numeric number of bands, Default: 1

with_nodata
PARAM_DESCRIPTION, Default: TRUE

tofile
logical if TRUE, the rasterStack is saved to a temporary file and the name of the temporary filename is returned instead than the stack itself, Default: FALSE

Value

either the "test" rasterStack, or a filename corresponding to it (if tofile = TRUE)

Examples

# NOT RUN {
if(interactive()){
 r <- build_testraster(100,100,10, tofile = F)
 }
# }