create_fishnet

funtion to create a raster and/or polygon fishnet of specified resolution over a a specified spatial extent. If both in_ext and crop_ext are declared, the fishnet is created according to in_ext (that is, the origin depends from in_ext), and then cropped on crop_ext (this allows for example to create a fishnet based on a given raster/resolution, and then to automatically crop it on a desired extgent without "moving" the corners of the cells ")

create_fishnet(in_obj, cellsize, crop_layer = NULL, out_raster = NULL,
  out_shape = NULL, overwrite = TRUE, return_sp = TRUE, pypath = NULL)

Arguments

in_obj
object of sf or raster* class from which the fishnet needs to be derived. The fishnet is built from the corner of in_obj, with a resolution equal to cellsize

cellsize
numeric Dimensions of the desired cells

crop_layer
(optional) object of Extent. If not null, the fishnet extent will be cropped on this extent, without "moving" the nodes of the grid. This is useful to crop a grid created on the basis of a different raster coordinates on top of a different raster, Default: NULL

out_raster
character (optional) name of the output raster file. If NULL the grid is NOT saved in a raster file Default: NULL

out_shape
character (optional) name of the output shape file. If NULL the grid is NOT saved in a shapefile, Default: NULL

overwrite
logical If TRUE, existing files are overwritten, Default: FALSE

return_sp
logical, If true, the grid is returned to the caller as a `SpatialGridDataFrame , Default: FALSE

pypath
path to gdal_polygonize.py. If NULL, the path is searched on the user's machine, Default: NULL

Value

if return_sp == TRUE, returns the grid as a SpatialGridDataFrame

Details

DETAILS

Examples

  # create a fishnet over an input raster file and save it as a shapefile
  library(raster)
  library(sprawl)

  file_in  <- raster::stack(system.file("extdata", "sprawl_EVItest.tif", package = "sprawl"))[[1]]
  inrast   <- raster::raster(file_in)
  cellsize <- raster::res(inrast)[1]

  out_file <- "/somedir/somefile.shp"
  fishnet  <- create_fishnet(inrast,
                             cellsize,
                             return_sp = TRUE
                             )

#> create_fishnet -> Writing Polygon Grid to: C:\Users\LB_LAP~1\AppData\Local\Temp\Rtmp0eUuAc\file32e474ae1717.shp - Please Wait !

#> Warning: running command '"python" C:\OSGeo4W64\bin\gdal_polygonize.py C:\Users\LB_LAP~1\AppData\Local\Temp\Rtmp0eUuAc\tempshp_32e436367a73.tif -f "ESRI Shapefile" C:\Users\LB_LAP~1\AppData\Local\Temp\Rtmp0eUuAc\file32e474ae1717.shp -fieldname id' had status 1

#> create_fishnet -> Reading Polygon Grid C:\Users\LB_LAP~1\AppData\Local\Temp\Rtmp0eUuAc\file32e474ae1717.shp to R - Please Wait !

#> Error in read_shape(out_shape): read_shape --> Input file doesn't exist on your system ! Aborting !

  plot(inrast)

#> Error in .plotraster2(x, col = col, maxpixels = maxpixels, add = add,     ext = ext, interpolate = interpolate, colNA = colNA, main = main,     addfun = addfun, facvar = facvar, alpha = alpha, npretty = npretty,     ...): no values associated with this RasterLayer

  plot(fishnet, add = TRUE, col = "transparent")

#> Error in plot(fishnet, add = TRUE, col = "transparent"): object 'fishnet' not found