crop a raster file/rast_file on a given extent

function to crop a raster file/rast_file on a given extent. The extent can be:

directly passed (as a sprawlext rast_file or numeric array);

derived from a different rast_file (raster of vector file/rast_file) passed to the function (see details)

crop_rast(rast_object, ext_object, out_type = "rastobject",
  out_filename = NULL, compress = "None", verbose = TRUE)

Arguments

rast_object
either an R object of class Raster, or a character string corresponding to a raster filename (with full path)

ext_object
either an object of class sprawlext, or an R object or filename from which an object of class sprawlext can be obtained (see sprawl::get_extent())

out_type
character indicating to which type of rast_file the input should be re-cropped.

"rastobject" return a raster rast_file accessing the saved cropped file

"rastfile" return the filename of the GTiff file corresponding to the cropped GTiff file. If out_filename == NULL, this corresponds to a file saved in R temporary folder.

"vrtfile" return the filename of the vrt file built for cropping the input raster (no saving to disk is performed) , Default: "rastobject"

out_filename
character filename to be used to save the cropped raster.

compress
character compression option to be used to saved the cropped raster ("None", "PACKBITS", "LZW", "DEFLATE), Default: "None"

verbose
logical if FALSE, suppress processing messages, Default: TRUE

Value

either a Raster object containing the cropped raster, or a gdal vrt or GTiff filename corresponding to it, depending on out_type.

Examples

 #EXAMPLE1