cz_points

FUNCTION_DESCRIPTION

cz_points(zone_object, in_rast, n_selbands, selbands, seldates, id_field,
  date_check, long, verbose, addfeat, addgeom, keep_null)

Arguments

zone_object
sf object on which features the zonal statistics has to be computed

in_rast
single- or multi-band rasterStack object containing values to be extracted and summarized

n_selbands
aa

selbands
2-element numeric array defining starting and ending raster bands to be processed (e.g., c(1, 10). Default: NULL, meaining that all bands will be processed

seldates
aaa

id_field
character (optional) field of the vector file to be used to identify the zones from which data is to be extracted. MUST correspond to a univoc attribute (i.e., no replicated values are allowed), If NULL (the default), zonal statistics are extracted on each row of the shapefile, and a new column named id_feat is used on the output to distinguish the zones, Default: NULL

date_check
zzz

long
PARAM_DESCRIPTION, Default: FALSE

verbose
PARAM_DESCRIPTION, Default: TRUE

addfeat
PARAM_DESCRIPTION, Default: TRUE

addgeom
PARAM_DESCRIPTION, Default: TRUE

keep_null
PARAM_DESCRIPTION, Default: FALSE

Value

OUTPUT_DESCRIPTION