dbReadSpatial

Import a spatial table / view / materialised view from a PostgreSQL database into a Spatial*DataFrame object.

dbReadSpatial(conn, name, pkey = NULL, ...)

Arguments

conn
DBIConnection object, as produced by DBI::dbConnect().

name
character the name of the spatial table / view / materialised view. If the table is not in the 'public' schema, use a two-length character vector in the form c('schema_name','table_name').

pkey
character (optional) the name of the primary key. This is needed only in the case of views or materialised views (the primary key of a table is automatically retrieved).

...
other options to be passed to....s

Value

An object of class SpatialPixelsDataFrame, SpatialLinesDataFrame or SpatialPolygonsDataFrame, depending on the specific geometry which is being read.

Details

The function reads a PostGIS table / view / materialised view from a PostgreSQL database, in which a spatial column (format geom) must be present, and import it in R as a *sp object SpatialPixelsDataFrame, SpatialLinesDataFrame or SpatialPolygonsDataFrame. The name of the geometric column is automatically retrieved, while the primary key is retrieved only reading a table (in case a view or materialised view is read, it must be specified with the parameter pkey).

Note

License: GPL 3.0

Examples