er_points

FUNCTION_DESCRIPTION

er_points(in_vect_zones, in_rast, n_selbands, selbands, seldates, id_field,
  date_check, long, verbose, addfeat, addgeom, keep_null)

Arguments

in_vect_zones
sf object on which features the zonal statistics has to be computed

in_rast
single- or multi-band rasterStack object containing values to be extracted and summarized

n_selbands
aa

selbands
2-element numeric array defining starting and ending raster bands to be processed (e.g., c(1, 10). Default: NULL, meaining that all bands will be processed

seldates
aaa

id_field
character (optional) field of the vector file to be used to identify the zones from which data is to be extracted, If NULL (or invalid) zonal statistics are extracted on each row of the shapefile, and a new column named id_feat is used on the output to distinguish the zones, Default: NULL

date_check
zzz

long
logical If TRUE, out$summ_data is returned in long format, Default: FALSE

verbose
logical If TRUE, messages concerning the processing status are shown in the console, Default: TRUE

addfeat
logical If TRUE, columns of the attribute table of the in_vect_zones layer are joined to results of the computation, Default: TRUE

addgeom
logical, If TRUE, the output sent out as an sf object, preserving the geometry of in_vect_zones. Note that this leads to dyuplication of geometries, and may be very slow for large datasets. Default: FALSE

keep_null
logical If TRUE, the output preserves features of in_vect_zones falling outside the extent of in_rast. Values for these features are set to NA, Default: FALSE

Value

OUTPUT_DESCRIPTION