extract raster values on features of a vector

Function used to extract values of a single- or multi- band raster for pixels corresponding to the features of a vector (Polygon, Point or Line)

extract_rast(in_rast, in_vect_zones, selbands = NULL, rastres = NULL,
  id_field = NULL, summ_data = TRUE, full_data = TRUE,
  comp_quant = FALSE, FUN = NULL, small = TRUE, na.rm = TRUE,
  maxchunk = 5e+07, long = FALSE, addfeat = TRUE, addgeom = TRUE,
  keep_null = FALSE, verbose = TRUE, ncores = NULL, mode = "std")

Arguments

in_rast
input raster. Can be wither:

A file name corresponding to a valid single- or multi- band raster (e.g. /my_folder/myfile.tif) All gdal raster formats are supported

An raster, rasterStack or rasterBrick object

in_vect_zones
input vector object containing zones from which data has to be extracted. Can be either:

A file name corresponding to a valid ESRI shapefile (e.g. /my_folder/myshape.shp)

An R +sp or sf object

selbands
2-element numeric array defining starting and ending raster bands to be processed (e.g., c(1, 10). Default: NULL, meaining that all bands will be processed

rastres
numeric, if not null, the input raster is resampled to rastres prior to data extraction using nearest neighbour resampling. This is useful in case the polygons are small with respect to in_rast resolution, Default: NULL (meaning no resampling is done)

id_field
character (optional) field of the vector file to be used to identify the zones from which data is to be extracted, If NULL (or invalid) zonal statistics are extracted on each row of the shapefile, and a new column named id_feat is used on the output to distinguish the zones, Default: NULL

summ_data
logical If TRUE, summary statistics of values belonging to each vector zone are returned in out$stats. Default statistics are average, median, standard deviation and variation coefficient. Quantiles of the distribution can be also returned if comp_quant == TRUE, Default: TRUE

full_data
logical If TRUE, values of all pixels belonging to each vector zone are returned in `out$alldata' (Setting this to FALSE can be a good idea for large datasets....), Default: TRUE

comp_quant
logical if TRUE, also quantiles of the distributions of values are computed for each zone and returned in out$ts_summ, Default: FALSE

FUN
character or function name, Default: NULL

small
logical if TRUE, values are returned also for small polygons not including any pixel centroids. Values are taken from all cells "touched" by the small polygon, Default: TRUE

na.rm
logical If TRUE, NA values are removed while computing statistics, Default: TRUE

maxchunk
Maximum chunk size (provisional), Default: 5e+06

long
logical If TRUE, out$summ_data is returned in long format, Default: FALSE

addfeat
logical If TRUE, columns of the attribute table of the in_vect_zones layer are joined to results of the computation, Default: TRUE

addgeom
logical, If TRUE, the output sent out as an sf object, preserving the geometry of in_vect_zones. Note that this leads to dyuplication of geometries, and may be very slow for large datasets. Default: FALSE

keep_null
logical If TRUE, the output preserves features of in_vect_zones falling outside the extent of in_rast. Values for these features are set to NA, Default: FALSE

verbose
logical If TRUE, messages concerning the processing status are shown in the console, Default: TRUE

ncores
numeric maximum number of cores to be used in the processin. If NULL, defaults to available cores - 2, but up to a maximum of 8. If user-provided is greter than available cores - 2 or greater than 8, ncores is re-set to the minimum between those two.

mode
character "std" or "velox", Default: "std"

Value

out_list list containing two tibbles: out_list$summ_data contains summary statitstics, while out_list$alldata contains the data of all pixels extracted (see examples).

Examples

library(sprawl)
library(sprawl.data)
library(raster)
options(tibble.width = Inf)
in_polys <- read_vect(system.file("extdata","lc_polys.shp", package = "sprawl.data"),
                       stringsAsFactors = T)
in_rast  <- raster::stack(system.file("extdata", "sprawl_EVItest.tif", package = "sprawl.data"))
in_rast  <- raster::setZ(in_rast, doytodate(seq(1,366, by = 8), year = 2013))
out      <- extract_rast(in_rast, in_polys, long = FALSE, verbose = FALSE)
head(out$stats)

#> Simple feature collection with 6 features and 13 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 12918420 ymin: 1752797 xmax: 12920610 ymax: 1755103
#> epsg (SRID):    NA
#> proj4string:    +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> # A tibble: 6 x 14
#>   id_feat band_n       date N_PIX      avg   med       sd   min   max    id  lc_type category sup_catego          geometry
#>     <dbl>  <dbl>     <date> <int>    <dbl> <dbl>    <dbl> <dbl> <dbl> <dbl>   <fctr>   <fctr>     <fctr> <S3: sfc_POLYGON>
#> 1       1      1 2013-01-01    61 4830.820  4747 540.6097  3731  5987     1 forest_1   Forest Vegetation <S3: sfc_POLYGON>
#> 2       1      2 2013-01-09    61 4605.131  4648 510.8426  3665  5547     1 forest_1   Forest Vegetation <S3: sfc_POLYGON>
#> 3       1      3 2013-01-17    61 4837.902  4834 557.4429  3640  5827     1 forest_1   Forest Vegetation <S3: sfc_POLYGON>
#> 4       1      4 2013-01-25    61 4484.984  4411 473.7452  3285  5275     1 forest_1   Forest Vegetation <S3: sfc_POLYGON>
#> 5       1      5 2013-02-02    61 4284.197  4237 498.7366  3469  5318     1 forest_1   Forest Vegetation <S3: sfc_POLYGON>
#> 6       1      6 2013-02-10    61 4374.525  4302 471.1914  3317  5363     1 forest_1   Forest Vegetation <S3: sfc_POLYGON>

head(out$alldata)

#> Simple feature collection with 6 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 12918200 ymin: 1754797 xmax: 12919360 ymax: 1754797
#> epsg (SRID):    NA
#> proj4string:    +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> # A tibble: 6 x 11
#>   mdxtnq band_n       date N_PIX     N value    id  lc_type category sup_catego        geometry
#>    <int>  <int>     <date> <int> <int> <dbl> <dbl>   <fctr>   <fctr>     <fctr> <S3: sfc_POINT>
#> 1      1      1 2013-01-01    61     1  4454     1 forest_1   Forest Vegetation <S3: sfc_POINT>
#> 2      1      1 2013-01-01    61     2  4260     1 forest_1   Forest Vegetation <S3: sfc_POINT>
#> 3      1      1 2013-01-01    61     3  4224     1 forest_1   Forest Vegetation <S3: sfc_POINT>
#> 4      1      1 2013-01-01    61     4  4344     1 forest_1   Forest Vegetation <S3: sfc_POINT>
#> 5      1      1 2013-01-01    61     5  4489     1 forest_1   Forest Vegetation <S3: sfc_POINT>
#> 6      1      1 2013-01-01    61     6  4970     1 forest_1   Forest Vegetation <S3: sfc_POINT>