Mask a raster file (or object) based on a vector file (or object)

Masks a raster file or object on the basis of a vector file or object. Pixels not covered by the vector features are set to NoData. If the input raster is multi-band, the mask is automatically applied to all bands. An optional buffer can be applied to the input vector to allow a more "lenient" masking, or to remove also the borders of the vector.

mask_rast(in_rast, mask_vect, crop = FALSE, buffer = NULL,
  out_nodata = "NoData", to_file = FALSE, out_rast = NULL,
  out_dt = "FLT4S", verbose = TRUE)

Arguments

in_rast
Raster file or object inheriting class raster to be masked

mask_vect
Vector file or object of class *sf or sp to be used as a mask

crop
logical if TRUE, in_rast is also cropped on the extent of mask_vect, Default: FALSE

buffer
numeric if not NULL, width of a buffer to be applied to mask_vect before masking in_rast. If negative, mask_vect is "reduced" prior to masking (see examples), Default: NULL

out_nodata
numeric value to be assigned to areas outside the mask, Default: 'NoData'

to_file
logical If TRUE, the output masked raster is save to file instead than sent back to the caller. In this case, the path to the saved file is returned instead

out_rast
character filename where the masked raster should be saved (ignorted if to_file == FALSE). If NULL while to_file == TRUE, the masked raster is saved on a temporaty file in the R temporary folder. The file is saved in TIFF format, with DAFLATE compression.

verbose
logical if TRUE, extended processing information is sent to the console in the form of messages

Value

object of class raster (if to_file == FALSE), or character string corresponding to the filename of the created raster (if to_file == TRUE)

Examples

libray(sprawl)

#> Error in libray(sprawl): could not find function "libray"

libray(sprawl.data)

#> Error in libray(sprawl.data): could not find function "libray"

library(raster)
in_polys <- read_shape(system.file("extdata","lc_polys.shp", package = "sprawl.data"), stringsAsFactors = T)
in_rast  <- raster::stack(system.file("extdata", "testrast.tif", package = "sprawl.data"))[[1]]

#> Error in .rasterObjectFromFile(x, objecttype = "RasterBrick", ...): provide a valid filename

in_polys <- sf::st_transform(in_polys, proj4string(in_rast))

#> Error in proj4string(in_rast): object 'in_rast' not found

masked   <- mask_rast(in_rast, in_polys, verbose = FALSE)

#> Error in inherits(object, "Spatial"): object 'in_rast' not found

par(mfrow=c(1,2))
plot(in_rast)

#> Error in plot(in_rast): object 'in_rast' not found

plot(masked)

#> Error in plot(masked): object 'masked' not found