comp_zonal

Function used to extract values of a single- or multi- band raster for pixels corresponding to the features of a vector (Polygon, Point or Line)

comp_zonal(in_rast, zone_object, selbands = NULL, mask_object = NULL,
  rastres = NULL, id_field = NULL, summ_data = TRUE, full_data = TRUE,
  comp_quant = FALSE, out_format = "dframe", small = TRUE, na.rm = TRUE,
  maxchunk = 5e+07, long = FALSE, addfeat = TRUE, addgeom = TRUE,
  keep_null = FALSE, verbose = TRUE, ncores = NULL)

Arguments

in_rast
input raster. Can be wither:

A file name corresponding to a valid single- or multi- band raster (e.g. /my_folder/myfile.tif) All gdal raster formats are supported

An R raster, 'rasterStackorrasterBrick` object

zone_object
input vector object. Can be either:

A file name corresponding to a valid ESRI shapefile (e.g. /my_folder/myshape.shp)

An R +sp or sf object

selbands
2-element numeric array defining starting and ending raster bands to be processed (e.g., c(1, 10). Default: NULL, meaining that all bands will be processed

mask_object
Not used for now - will allow to mask the raster before extraction of values, Default: NULL

rastres
PARAM_DESCRIPTION, Default: NULL

id_field
character (optional) field of the vector file to be used to identify the zones from which data is to be extracted. MUST correspond to a univoc attribute (i.e., no replicated values are allowed), If NULL (the default), zonal statistics are extracted on each row of the shapefile, and a new column named id_feat is used on the output to distinguish the zones, Default: NULL

summ_data
logical If TRUE, summary statistics of values belonging to each vector zone are returned in out$ts_summ~. Default statistics are average, median, standard deviation and variation coefficient. Quantiles of the distribution can be also returned ifcomp_quant == TRUE`, Default: TRUE

full_data
logical If TRUE, values of all pixels belonging to each vector zone are returned in `out$ts_full', Default: TRUE

comp_quant
logical if TRUE, also quantiles of the distributions of values are computed for each zone and returned in out$ts_summ, Default: FALSE

out_format
character either "dframe" or "xts" Define the output format, Default: 'dframe'

small
logical if TRUE, values are returned also for small polygons not including any pixel centroids, Default: TRUE

na.rm
PARAM_DESCRIPTION, Default: TRUE

maxchunk
PARAM_DESCRIPTION, Default: 5e+07

long
PARAM_DESCRIPTION, Default: FALSE

addfeat
PARAM_DESCRIPTION, Default: TRUE

addgeom
PARAM_DESCRIPTION, Default: TRUE

keep_null
PARAM_DESCRIPTION, Default: FALSE

verbose
PARAM_DESCRIPTION, Default: TRUE

ncores
PARAM_DESCRIPTION

Value

OUTPUT_DESCRIPTION

Details

DETAILS

Examples