extract raster values on features of a vector

Function used to extract values of a single- or multi- band raster for pixels corresponding to the features of a vector (Polygon, Point or Line)

extract_rast(in_rast, in_vect, selbands = NULL, rastres = NULL,
  id_field = NULL, summ_data = TRUE, full_data = TRUE,
  comp_quant = FALSE, long = FALSE, FUN = NULL, small = TRUE,
  na.rm = TRUE, maxchunk = 5e+07, addfeat = TRUE, addgeom = TRUE,
  keep_null = FALSE, verbose = TRUE, ncores = NULL)

Arguments

in_rast
input raster. Can be wither:

A file name corresponding to a valid single- or multi- band raster (e.g. /my_folder/myfile.tif) All gdal raster formats are supported

An raster, rasterStack or rasterBrick object

in_vect
input vector object containing zones from which data has to be extracted. Can be either:

A file name corresponding to a valid ESRI shapefile (e.g. /my_folder/myshape.shp)

An R +sp or sf object

selbands
2-element numeric array defining starting and ending raster bands to be processed (e.g., c(1, 10). Default: NULL, meaining that all bands will be processed

rastres
numeric, if not null, the input raster is resampled to rastres prior to data extraction using nearest neighbour resampling. This is useful in case the polygons are small with respect to in_rast resolution, Default: NULL (meaning no resampling is done)

id_field
character (optional) field of the vector file to be used to identify the zones from which data is to be extracted, If NULL (or invalid) zonal statistics are extracted on each row of the shapefile, and a new column named id_feat is used on the output to distinguish the zones, Default: NULL

summ_data
logical If TRUE, summary statistics of values belonging to each vector zone are returned in out$stats. Default statistics are average, median, standard deviation and variation coefficient. Quantiles of the distribution can be also returned if comp_quant == TRUE, Default: TRUE

full_data
logical If TRUE, values of all pixels belonging to each vector zone are returned in `out$alldata' (Setting this to FALSE can be a good idea for large datasets....), Default: TRUE

comp_quant
logical if TRUE, also quantiles of the distributions of values are computed for each zone and returned in out$ts_summ, Default: FALSE

long
logical if TRUE, extraction on points provides data in long form

FUN
character or function name, Default: NULL

small
logical if TRUE, values are returned also for small polygons not including any pixel centroids. Values are taken from all cells "touched" by the small polygon, Default: TRUE

na.rm
logical If TRUE, NA values are removed while computing statistics, Default: TRUE

maxchunk
Maximum chunk size (provisional), Default: 5e+06

addfeat
logical If TRUE, columns of the attribute table of the in_vect layer are joined to results of the computation, Default: TRUE

addgeom
logical, If TRUE, the output sent out as an sf object, preserving the geometry of in_vect. Note that this leads to duplication of geometries, and may be very slow for large datasets. Default: FALSE

keep_null
logical If TRUE, the output preserves features of in_vect falling outside the extent of in_rast. Values for these features are set to NA, Default: FALSE

verbose
logical If TRUE, messages concerning the processing status are shown in the console, Default: TRUE

ncores
numeric maximum number of cores to be used in the processin. If NULL, defaults to available cores - 2, but up to a maximum of 8. If user-provided is greater than available cores - 2 or greater than 8, ncores is re-set to the minimum between those two.

Value

out_list list containing two tibbles: out_list$summ_data contains summary statitstics, while out_list$alldata contains the data of all pixels extracted (see examples).

Examples

library(sprawl)
library(sprawl.data)
library(raster)
library(tibble)
in_polys <- read_vect(system.file("extdata","lc_polys.shp", package = "sprawl.data"),
                       stringsAsFactors = T)
in_rast  <- raster::stack(system.file("extdata", "sprawl_EVItest.tif", package = "sprawl.data"))
in_rast  <- raster::setZ(in_rast, doytodate(seq(1,366, by = 8), year = 2013))
out      <- extract_rast(in_rast, in_polys, verbose = FALSE)

#> extract_rast --> Extracting: in_rastdata on zones of : in_polys

#> Error in {    if (er_opts$verbose) {        message("extract_rast--> Extracting data from ", ifelse(date_check,             " dates", "bands"), " - Please wait !")    }    all_data <- list()    coords <- list()    stat_data <- list()    temp_outdata <- list()    selband <- seldates[band]    chunk_n_all <- 1    chunk_n_summ <- 1    start_cell <- 1    in_band <- in_rast[[selbands[band]]]    tempvrt <- tempfile(fileext = ".vrt")    if (in_band@file@name == "") {        temprastfile <- tempfile(fileext = ".tif")        writeRaster(in_band, filename = temprastfile, options = c("COMPRESS=DEFLATE",             "PREDICTOR=3"), overwrite = TRUE)        in_band <- raster(temprastfile)    }    buildvrt_string <- paste("-te ", paste(te, collapse = " "),         "-b ", band, tempvrt, in_band@file@name, " ")    system2(file.path(find_gdal(), "gdalbuildvrt"), args = buildvrt_string,         stdout = NULL)    in_band <- raster::raster(tempvrt)    if (supersample) {        temprast <- tempfile(fileext = ".tif")        raster::writeRaster(in_band, temprast)        tempsuper <- tempfile(fileext = ".tif")        in_band <- gdalUtils::gdalwarp(temprast, tempsuper, tr = er_opts$rastres,             te = raster::extent(in_band)[c(1, 3, 2, 4)], output_Raster = TRUE,             multi = TRUE)        in_band <- velox::velox(in_band)        in_band$crop(sp_polys)    }    if (n_chunks > 1) {        bboxes <- in_vect_crop[c("mdxtnq", "geometry")] %>% data.table::data.table()        bboxes = bboxes[, list(min_y = sf::st_bbox(geometry)[2],             max_y = sf::st_bbox(geometry)[4]), by = "mdxtnq"]    }    else {        bboxes <- in_vect_crop[c("mdxtnq", "geometry")]    }    for (chunk in seq_len(n_chunks)) {        if (er_opts$verbose)             message("Working on chunk: ", chunk, " of: ", n_chunks,                 " of band: ", selband)        startrow <- ifelse(chunk == 1, 1, 1 + (chunk - 1) * ceiling(nrows/n_chunks))        chunkrows <- ifelse(chunk != n_chunks, ceiling(nrows/n_chunks),             (1 + nrows - startrow))        endrow <- startrow + chunkrows - 1        ncells <- ncols * chunkrows        end_cell <- start_cell + ncells - 1        out_data <- data.table::data.table(value = as.numeric(raster::getValues(in_band,             startrow, chunkrows)), cell = seq(start_cell, end_cell),             mdxtnq = as.numeric(raster::getValues(rast_zoneobject,                 startrow, chunkrows)), key = "mdxtnq")        out_data <- out_data[mdxtnq != 0]        ext_chunk <- data.frame(x_min = raster::extent(in_band)[1],             x_max = raster::extent(in_band)[2], y_min = raster::yFromRow(in_band,                 endrow), y_max = raster::yFromRow(in_band, 1))        if (dim(out_data)[1] > 0) {            if (er_opts$full_data) {                if (er_opts$addgeom) {                  temp_velox <- velox::velox(matrix(nrow = chunkrows,                     ncol = ncols), ext = as.numeric(ext_chunk),                     res = er_opts$rastres, crs = sp::proj4string(in_band))                  coords <- temp_velox$getCoordinates()[(out_data$cell -                     start_cell), ]                  out_data <- out_data[, `:=`(c("cell", "x_coord",                     "y_coord"), list(NULL, coords[, 1], coords[,                     2]))]                }                else {                  out_data <- out_data[, `:=`(cell, NULL)]                }                all_data[[chunk_n_all]] <- out_data                chunk_n_all <- chunk_n_all + 1            }            if (er_opts$summ_data) {                tot_ext_y <- data.frame(y_min = (raster::yFromRow(in_band,                   endrow) - er_opts$rastres[1]/2), y_max = (raster::yFromRow(in_band,                   1) + er_opts$rastres[1]/2))                temp_outdata <- data.table::rbindlist(list(temp_outdata,                   out_data))                if (n_chunks > 1) {                  complete_polys <- bboxes[min_y >= tot_ext_y$y_min]                }                else {                  complete_polys <- bboxes                }                if (length(complete_polys$mdxtnq) != 0) {                  data_for_summary <- subset(temp_outdata, mdxtnq %in%                     unique(complete_polys$mdxtnq)) %>% data.table::setkey("mdxtnq")                  stat_data[[chunk_n_summ]] <- summarize_data(data_for_summary,                     "mdxtnq", er_opts$comp_quant, er_opts$FUN,                     selbands[band], selband)                  temp_outdata <- temp_outdata[!(mdxtnq %in%                     unique(complete_polys$mdxtnq))]                  chunk_n_summ <- chunk_n_summ + 1                }            }        }        start_cell <- end_cell + 1    }    if (er_opts$full_data) {        all_data <- data.table::rbindlist(all_data) %>% data.table::setkey("mdxtnq")    }    if (er_opts$summ_data) {        stat_data <- data.table::rbindlist(stat_data) %>% data.table::setkey("mdxtnq")    }    if (er_opts$small & length(unique(stat_data$mdxtnq) != length(unique(in_vect_crop$mdxtnq)))) {        miss_feat <- setdiff(unique(in_vect_crop$mdxtnq), unique(stat_data$mdxtnq))        for (mfeat in miss_feat) {            poly_miss <- in_vect_crop %>% dplyr::filter(mdxtnq ==                 mfeat) %>% sf::st_as_sf() %>% as("Spatial")            data_feat <- raster::extract(in_rast[[band]], poly_miss,                 small = TRUE, method = "simple", df = TRUE, cellnumbers = TRUE)            cell <- data_feat[, 2]            miss_feat_data <- data.table::data.table(value = data_feat[,                 3], cell = cell, mdxtnq = mfeat)            if (er_opts$addgeom) {                coords <- raster::xyFromCell(in_band, cell)                miss_feat_data <- miss_feat_data[, `:=`(c("cell",                   "x_coord", "y_coord"), list(NULL, coords[,                   1], coords[, 2]))]            }            else {                miss_feat_data <- miss_feat_data[, `:=`(cell,                   NULL)]            }            if (er_opts$full_data) {                all_data <- rbind(all_data, miss_feat_data)            }            if (er_opts$summ_data) {                miss_feat_stats <- summarize_data(miss_feat_data,                   "mdxtnq", er_opts$comp_quant, er_opts$FUN,                   selbands[band], selband)                stat_data <- rbind(stat_data, miss_feat_stats)            }        }    }    if (er_opts$full_data) {        all_data <- all_data[, `:=`(c("band_n", "date", "N_PIX",             "N"), list(band, seldates[band], .N, seq_len(.N))),             by = mdxtnq]    }    if (er_opts$verbose) {        Sys.sleep(0.001)        utils::setTxtProgressBar(pb, band)    }    out <- list(alldata = all_data, stats = stat_data)    return(out)}: task 2 failed - "could not find function "summarize_data""

as_tibble(out$stats)

#> Error in as_tibble(out$stats): object 'out' not found

as_tibble(out$alldata)

#> Error in as_tibble(out$alldata): object 'out' not found