extract raster values on features of a vector

Function used to extract values of a single- or multi- band raster for pixels corresponding to the features of a vector (Polygon, Point or Line)

extract_rast(in_rast, in_vect, rast_type = "continuous", selbands = NULL,
  rastres = NULL, id_field = NULL, summ_data = TRUE, full_data = TRUE,
  comp_quant = FALSE, comp_freq = FALSE, long_format = TRUE, FUN = NULL,
  small = TRUE, na.rm = TRUE, maxchunk = 5e+07, join_feat_tbl = TRUE,
  join_geom = TRUE, keep_null = FALSE, verbose = TRUE, ncores = NULL)

Arguments

in_rast
input raster. Can be either:

A file name corresponding to a valid single- or multi- band raster (e.g. /my_folder/myfile.tif) All gdal raster formats are supported

An raster, rasterStack or rasterBrick object

in_vect
input vector object containing zones from which data has to be extracted. Can be either:

A file name corresponding to a valid ESRI shapefile (e.g. /my_folder/myshape.shp)

An R +sp or sf object

rast_type
character ("continuous" | "categorical) specifies if the values passed in in_data represent a continuous or categorical variable. (see @description)

selbands
2-element numeric array defining starting and ending raster bands to be processed (e.g., c(1, 10). Default: NULL, meaning that all bands will be processed

rastres
numeric(2), if not null, the input raster is resampled to rastres prior to data extraction using nearest neighbour resampling. This is useful in case the polygons are small with respect to in_rast resolution, Default: NULL (meaning no resampling is done)

id_field
character (optional) field of the vector file to be used to identify the zones from which data is to be extracted, If NULL (or invalid) zonal statistics are extracted on each row of the shapefile, and a new column named id_feat is used on the output to distinguish the zones, Default: NULL

summ_data
logical If TRUE, summary statistics of values belonging to each vector zone are returned in out$stats. Default statistics are average, median, standard deviation and variation coefficient. Quantiles of the distribution can be also returned if comp_quant == TRUE, Default: TRUE

full_data
logical If TRUE, values of all pixels belonging to each vector zone are returned in out$alldata (Setting this to FALSE can be a good idea for large datasets), Default: TRUE

comp_quant
logical if TRUE, also quantiles of the distributions of values are computed for each zone and returned in out$ts_summ, Default: FALSE (Ignored if rast_type == "categorical")

comp_freq
logical if TRUE, also frequencies of the the different classes of a categorical raster present in each polygon are returned, in out$ts_summ Default = FALSE (Ignored if rast_type == "continuous")

long_format
logical if TRUE, extraction on points provides data in long format (i.e., only one "value" column, with an "id" column) allowing to select specific points, Default: TRUE

FUN
character or function name, Default: NULL

small
logical if TRUE, values are returned also for small polygons not including any pixel centroids. Values are taken from all cells "touched" by the small polygon, Default: TRUE

na.rm
logical If TRUE, NA values are removed while computing statistics, Default: TRUE

maxchunk
Maximum chunk size (provisional), Default: 5e+06

join_feat_tbl
logical If TRUE, columns of the attribute table of the in_vect layer are joined to results of the computation, Default: TRUE

join_geom
logical, If TRUE, the output sent out as an sf object, preserving the geometry of in_vect. Note that this leads to duplication of geometries, and may be very slow for large datasets. Default: FALSE

keep_null
logical If TRUE, the output preserves features of in_vect falling outside the extent of in_rast. Values for these features are set to NA, Default: FALSE

verbose
logical If TRUE, messages concerning the processing status are shown in the console, Default: TRUE

ncores
numeric maximum number of cores to be used in the processing. If NULL, defaults to available cores - 2, but up to a maximum of 8. If user-provided is greater than available cores - 2 or greater than 8, ncores is re-set to the minimum between those two.

Value

out_list list containing two tibbles: out_list$summ_data contains summary statistics, while out_list$alldata contains the data of all pixels extracted (see examples).

Examples

library(sprawl)
library(sprawl.data)
library(raster)
library(tibble)
in_polys <- read_vect(system.file("extdata/shapes","lc_polys.shp",
                      package = "sprawl.data"), stringsAsFactors = T)
in_rast  <- raster::stack(system.file("extdata/MODIS_test", "EVIts_test.tif",
                          package = "sprawl.data"))
in_rast  <- raster::setZ(in_rast, doytodate(seq(1,366, by = 8), year = 2013))

out      <- extract_rast(in_rast, in_polys, verbose = FALSE)

# Statistics for the different polygons

as_tibble(out$stats)

#> Simple feature collection with 552 features and 15 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 12897990 ymin: 1685995 xmax: 12971590 ymax: 1780601
#> epsg (SRID):    NA
#> proj4string:    +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> # A tibble: 552 x 16
#>    id_feat band_n       date        area n_pix n_pix_val      avg   med
#>      <dbl>  <dbl>     <date> <S3: units> <int>     <int>    <dbl> <dbl>
#>  1       1      1 2013-01-01 3400057 m^2    61        61 4830.820  4747
#>  2       1      2 2013-01-09 3400057 m^2    61        61 4605.131  4648
#>  3       1      3 2013-01-17 3400057 m^2    61        61 4837.902  4834
#>  4       1      4 2013-01-25 3400057 m^2    61        61 4484.984  4411
#>  5       1      5 2013-02-02 3400057 m^2    61        61 4284.197  4237
#>  6       1      6 2013-02-10 3400057 m^2    61        61 4374.525  4302
#>  7       1      7 2013-02-18 3400057 m^2    61        61 4113.246  4123
#>  8       1      8 2013-02-26 3400057 m^2    61        61 4050.689  4164
#>  9       1      9 2013-03-06 3400057 m^2    61        61 3937.738  3917
#> 10       1     10 2013-03-14 3400057 m^2    61        61 4281.672  4282
#> # ... with 542 more rows, and 8 more variables: sd <dbl>, min <dbl>, max <dbl>,
#> #   id <dbl>, lc_type <fctr>, category <fctr>, sup_catego <fctr>,
#> #   geometry <simple_feature>

# Data from all pixels, grouped by polygon

as_tibble(out$alldata)

#> Simple feature collection with 147384 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 12898050 ymin: 1685995 xmax: 12971710 ymax: 1780279
#> epsg (SRID):    NA
#> proj4string:    +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> # A tibble: 147,384 x 11
#>    id_feat band_n       date n_pix     N value    id  lc_type category
#>      <int>  <int>     <date> <int> <int> <dbl> <dbl>   <fctr>   <fctr>
#>  1       1      1 2013-01-01    61     1  4454     1 forest_1   Forest
#>  2       1      1 2013-01-01    61     2  4260     1 forest_1   Forest
#>  3       1      1 2013-01-01    61     3  4224     1 forest_1   Forest
#>  4       1      1 2013-01-01    61     4  4344     1 forest_1   Forest
#>  5       1      1 2013-01-01    61     5  4489     1 forest_1   Forest
#>  6       1      1 2013-01-01    61     6  4970     1 forest_1   Forest
#>  7       1      1 2013-01-01    61     7  3987     1 forest_1   Forest
#>  8       1      1 2013-01-01    61     8  3904     1 forest_1   Forest
#>  9       1      1 2013-01-01    61     9  3904     1 forest_1   Forest
#> 10       1      1 2013-01-01    61    10  4768     1 forest_1   Forest
#> # ... with 147,374 more rows, and 2 more variables: sup_catego <fctr>,
#> #   geometry <simple_feature>