create list of valid fill palettes

helper function used to create a data.frame containing info on valid fill palettes for plot_rast, plot_rast_gg and plot_vect

fillpals()

Value

a data.frame containing info on valid fill palettes that can used in plot_rast, plot_rast_gg and plot_vect

Details

DETAILS

Examples

# Print the list of valid fill palettes used in `plot_rast`, `plot_rast_gg`
# and `plot_vect`
fillpals()

#>        name     source category cont_qual maxcolors colorblind
#> 1      BrBG     brewer      div      cont        11       TRUE
#> 2      PiYG     brewer      div      cont        11       TRUE
#> 3      PRGn     brewer      div      cont        11       TRUE
#> 4      PuOr     brewer      div      cont        11       TRUE
#> 5      RdBu     brewer      div      cont        11       TRUE
#> 6      RdGy     brewer      div      cont        11      FALSE
#> 7    RdYlBu     brewer      div      cont        11       TRUE
#> 8    RdYlGn     brewer      div      cont        11      FALSE
#> 9  Spectral     brewer      div      cont        11      FALSE
#> 10    Blues     brewer      seq      cont         9       TRUE
#> 11     BuGn     brewer      seq      cont         9       TRUE
#> 12     BuPu     brewer      seq      cont         9       TRUE
#> 13     GnBu     brewer      seq      cont         9       TRUE
#> 14   Greens     brewer      seq      cont         9       TRUE
#> 15    Greys     brewer      seq      cont         9       TRUE
#> 16  Oranges     brewer      seq      cont         9       TRUE
#> 17     OrRd     brewer      seq      cont         9       TRUE
#> 18     PuBu     brewer      seq      cont         9       TRUE
#> 19   PuBuGn     brewer      seq      cont         9       TRUE
#> 20     PuRd     brewer      seq      cont         9       TRUE
#> 21  Purples     brewer      seq      cont         9       TRUE
#> 22     RdPu     brewer      seq      cont         9       TRUE
#> 23     Reds     brewer      seq      cont         9       TRUE
#> 24     YlGn     brewer      seq      cont         9       TRUE
#> 25   YlGnBu     brewer      seq      cont         9       TRUE
#> 26   YlOrBr     brewer      seq      cont         9       TRUE
#> 27   YlOrRd     brewer      seq      cont         9       TRUE
#> 28    magma virdisLite      seq      cont       256       TRUE
#> 29  inferno virdisLite      seq      cont       256       TRUE
#> 30   plasma virdisLite      seq      cont       256       TRUE
#> 31  viridis virdisLite      seq      cont       256       TRUE
#> 32  cividis virdisLite      seq      cont       256       TRUE
#> 33   Accent     brewer     qual      qual         8      FALSE
#> 34    Dark2     brewer     qual      qual         8       TRUE
#> 35   Paired     brewer     qual      qual        12       TRUE
#> 36  Pastel1     brewer     qual      qual         9      FALSE
#> 37  Pastel2     brewer     qual      qual         8      FALSE
#> 38     Set1     brewer     qual      qual         9      FALSE
#> 39     Set2     brewer     qual      qual         8       TRUE
#> 40     Set3     brewer     qual      qual        12      FALSE
#> 41      hue     ggplot     qual      qual      1000      FALSE
#> 42   manual       user     qual      qual        NA         NA